Format arrays as []
